se actualizó Home.jsx para que se conecte a la BD para mostrar los productos, para elo se crearon loc componentes ProductOnSale.jsx y ProductOnSaleList.jsx
